requirements.txt
pandas
numpy
scipy
pypfopt
yfinance
PyYAML
openpyxl
xlsxwriter
python-dateutil